Set container port to 8053 by default in Helm chart (#9)
* (Helm) Allowing to set containerPort

The `spec.template.spec.containers[0].ports[0].containerPort` parameter
indicate the port the application is listening to inside the container.
Used with `env.DOH_SERVER_LISTEN`, it allows to set the listening port
of the application (previously hardcoded to 80).

* (Helm) Forcing app to start on port 8053

By default, the container does not start as root (cf #3) and the pod was
binding to port 80 which prevent it to start on this error:

    2021/12/29 01:37:10 listen tcp :80: bind: permission denied

Port below 1024 are reserved for privileged services and not
available for non root users. By setting it to 8053 as default,
the container will start smoothly on a non reserved port.

To manually set another port, both variables `env.DOH_SERVER_LISTEN` and
`containerPort` must be updated.